:sparkles: [김채영] 휴가 종류 목록 & 수정 완료

+++ src/main/java/com/fiveLink/linkOffice/vacation/domain/VacationTypeDto.java
@@ -16,7 +16,7 @@ import java.util.Map;
 public class VacationTypeDto {
     private Long vacation_type_no;
     private String vacation_type_name;
-    private int vacation_type_calculate;
+    private double vacation_type_calculate;
     @Builder.Default
     private Map<String, String> vacationTypesData = new HashMap<>();
